Remove all trace of coord in model, and switch from int to double in shape

+++ src/main/java/application/model/shape/Shape.java
@@ -4,23 +4,23 @@ import application.exceptions.UnsupportedOperationException;
 import application.view.View;
 
 public interface Shape {
-    void add(Shape shape) throws UnsupportedOperationException;
+    void add(Shape shape)    throws UnsupportedOperationException;
     void remove(Shape shape) throws UnsupportedOperationException;
 
     void draw(View view);
 
-    int getWidth();
-    int getHeight();
+    double getWidth();
+    double getHeight();
 
-    int getMinX();
-    int getMinY();
-    int getMaxX();
-    int getMaxY();
+    double getMinX();
+    double getMinY();
+    double getMaxX();
+    double getMaxY();
 
-    boolean isIn(int x, int y);
+    boolean isIn(double x, double y);
     boolean intersect(Rectangle rectangle);
-    void moveTo(int x, int y);
-    void translate(int dx, int dy);
+    void moveTo(double x, double y);
+    void translate(double dx, double dy);
 
     void resize(double factor);
     void resize(Shape containerShape, double factor);
